🛠Rename: move Search bar file
